done breaking out the SDK into modules
