+++ packages/studiocms/src/sdk/sdkCore.ts
@@ -1,13 +1,4 @@
-import { and, asc, desc, eq } from 'astro:db';
-import { createTwoFilesPatch } from 'diff';
-import { type Diff2HtmlConfig, html } from 'diff2html';
 import { Effect } from 'effect';
-import {
-	CMSNotificationSettingsId,
-	CMSSiteConfigId,
-	GhostUserDefaults,
-	NotificationSettingsDefaults,
-} from '../consts.js';
 import {
 	AstroDB,
 	SDKCore_Collectors,
@@ -16,26 +7,18 @@ import {
 	SDKCore_Parsers,
 	SDKCore_Users,
 } from './effect/index.js';
-import { SDKCoreError, StudioCMS_SDK_Error } from './errors.js';
+import { SDKCoreError } from './errors.js';
+import { SDKCore_AUTH } from './modules/auth.js';
 import { SDKCore_CLEAR } from './modules/clear.js';
 import { SDKCore_DELETE } from './modules/delete.js';
+import { SDKCore_DiffTracking } from './modules/diffTracking.js';
 import { SDKCore_GET } from './modules/get.js';
+import { SDKCore_INIT } from './modules/init.js';
+import { SDKCore_NotificationSettings } from './modules/notificationSettings.js';
 import { SDKCore_POST } from './modules/post.js';
+import { SDKCore_ResetTokenBucket } from './modules/resetTokenBucket.js';
 import { SDKCore_REST_API } from './modules/rest_api.js';
 import { SDKCore_UPDATE } from './modules/update.js';
-import {
-	tsDiffTracking,
-	tsEmailVerificationTokens,
-	tsNotificationSettings,
-	tsOAuthAccounts,
-	tsPageContent,
-	tsPageData,
-	tsPermissions,
-	tsSessionTable,
-	tsSiteConfig,
-	tsUserResetTokens,
-	tsUsers,
-} from './tables.js';
 import type {
 	CacheMap,
 	CombinedPageData,
@@ -47,16 +30,9 @@ import type {
 	PageDataCacheObject,
 	SiteConfigCacheObject,
 	VersionCacheObject,
-	tsNotificationSettingsInsert,
-	tsOAuthAccountsSelect,
 	tsPageDataSelect,
 	tsPermissionsSelect,
-	tsSessionTableInsert,
-	tsSiteConfigSelect,
-	tsUserResetTokensSelect,
-	tsUsersInsert,
 	tsUsersSelect,
-	tsUsersUpdate,
 } from './types/index.js';
 import { CacheContext, _ClearUnknownError, _clearLibSQLError } from './utils.js';
 
@@ -92,7 +68,12 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 		SDKCore_POST.Default,
 		SDKCore_REST_API.Default,
 		SDKCore_GET.Default,
-		AstroDB.Default
+		SDKCore_ResetTokenBucket.Default,
+		SDKCore_DiffTracking.Default,
+		SDKCore_NotificationSettings.Default,
+		SDKCore_AUTH.Default,
+		SDKCore_INIT.Default,
+		AstroDB.Default,
 	],
 	effect: Effect.gen(function* () {
 		// Get Services
@@ -108,7 +89,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 				getAvailableFolders,
 			},
 			{ generateRandomIDNumber, generateRandomPassword, generateToken, testToken },
-			{ parseIdNumberArray, parseIdStringArray, fixDiff },
+			{ parseIdNumberArray, parseIdStringArray },
 			{ combineRanks, verifyRank, clearUserReferences },
 			{ collectCategories, collectTags, collectPageData, collectUserData },
 			dbService,
@@ -118,6 +99,11 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 			POST,
 			REST_API,
 			GET,
+			resetTokenBucket,
+			diffTracking,
+			notificationSettings,
+			AUTH,
+			INIT,
 		] = yield* Effect.all([
 			SDKCore_FolderTree,
 			SDKCore_Generators,
@@ -131,1035 +117,15 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 			SDKCore_POST,
 			SDKCore_REST_API,
 			SDKCore_GET,
+			SDKCore_ResetTokenBucket,
+			SDKCore_DiffTracking,
+			SDKCore_NotificationSettings,
+			SDKCore_AUTH,
+			SDKCore_INIT,
 		]);
 
 		// Breakout service functions that need to be returned in this.
 		const { db } = dbService;
-
-		const resetTokenBucket = {
-			new: (userId: string): Effect.Effect<tsUserResetTokensSelect, SDKCoreError, never> =>
-				Effect.gen(function* () {
-					const token = yield* generateToken(userId);
-
-					return yield* dbService.execute((db) =>
-						db
-							.insert(tsUserResetTokens)
-							// @ts-expect-error Drizzle... removed this from the type?
-							.values({ id: crypto.randomUUID(), userId, token })
-							.returning()
-							.get()
-					);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`resetTokenBucket New Error: ${cause}`),
-								})
-							),
-					})
-				),
-			delete: (userId: string): Effect.Effect<void, SDKCoreError, never> =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsUserResetTokens).where(eq(tsUserResetTokens.userId, userId))
-					);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`resetTokenBucket Delete Error: ${cause}`),
-								})
-							),
-					})
-				),
-			check: (token: string): Effect.Effect<boolean, SDKCoreError, never> =>
-				Effect.gen(function* () {
-					const { isValid, userId } = yield* testToken(token);
-
-					if (!isValid) return false;
-					if (!userId) return false;
-
-					const resetToken = yield* dbService.execute((db) =>
-						db.select().from(tsUserResetTokens).where(eq(tsUserResetTokens.userId, userId))
-					);
-
-					if (!resetToken || resetToken.length === 0) return false;
-
-					return !!resetToken.find((t) => t.token === token);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`resetTokenBucket Check Error: ${cause}`),
-								})
-							),
-					})
-				),
-		};
-
-		const checkDiffsLengthAndRemoveOldestIfToLong = (
-			pageId: string,
-			length: number
-		): Effect.Effect<void, SDKCoreError, never> =>
-			Effect.gen(function* () {
-				const diffs = yield* dbService.execute((db) =>
-					db
-						.select()
-						.from(tsDiffTracking)
-						.where(eq(tsDiffTracking.pageId, pageId))
-						.orderBy(asc(tsDiffTracking.timestamp))
-				);
-
-				if (diffs.length > length) {
-					const oldestDiff = diffs[0];
-
-					yield* dbService.execute((db) =>
-						db.delete(tsDiffTracking).where(eq(tsDiffTracking.id, oldestDiff.id))
-					);
-				}
-			}).pipe(
-				Effect.catchTags({
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						Effect.fail(
-							new SDKCoreError({
-								type: 'LibSQLDatabaseError',
-								cause: new StudioCMS_SDK_Error(
-									`checkDiffsLengthAndRemoveOldestIfToLong Error: ${cause}`
-								),
-							})
-						),
-				})
-			);
-
-		const diffTracking = {
-			insert: (
-				userId: string,
-				pageId: string,
-				data: {
-					content: {
-						start: string;
-						end: string;
-					};
-					metaData: {
-						start: Partial<tsPageDataSelect>;
-						end: Partial<tsPageDataSelect>;
-					};
-				},
-				diffLength: number
-			) =>
-				Effect.gen(function* () {
-					const diff = yield* Effect.try({
-						try: () =>
-							createTwoFilesPatch('Content', 'Content', data.content.start, data.content.end),
-						catch: (error) =>
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error(
-									`diffTracking.insert:createTwoFilesPatch Error: ${error}`
-								),
-							}),
-					});
-
-					yield* checkDiffsLengthAndRemoveOldestIfToLong(pageId, diffLength);
-
-					const inputted = yield* dbService.execute((db) =>
-						db
-							.insert(tsDiffTracking)
-							.values({
-								// @ts-expect-error Drizzle... removed this from the type?
-								id: crypto.randomUUID(),
-								userId,
-								pageId,
-								diff,
-								timestamp: new Date(),
-								pageContentStart: data.content.start,
-								pageMetaData: JSON.stringify(data.metaData),
-							})
-							.returning()
-							.get()
-					);
-
-					return yield* fixDiff(inputted);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`diffTracking.insert Error: ${cause}`),
-								})
-							),
-					})
-				),
-			clear: (pageId: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsDiffTracking).where(eq(tsDiffTracking.pageId, pageId))
-					);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`diffTracking.clear Error: ${cause}`),
-								})
-							),
-					})
-				),
-			get: {
-				byPageId: {
-					all: (pageId: string) =>
-						Effect.gen(function* () {
-							const items = yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsDiffTracking)
-									.where(eq(tsDiffTracking.pageId, pageId))
-									.orderBy(desc(tsDiffTracking.timestamp))
-							);
-
-							return yield* fixDiff(items);
-						}).pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`diffTracking.get.byPageId.all Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-					latest: (pageId: string, count: number) =>
-						Effect.gen(function* () {
-							const items = yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsDiffTracking)
-									.where(eq(tsDiffTracking.pageId, pageId))
-									.orderBy(desc(tsDiffTracking.timestamp))
-							);
-
-							const split = items.slice(0, count);
-
-							return yield* fixDiff(split);
-						}).pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`diffTracking.get.byPageId.latest Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-				},
-				byUserId: {
-					all: (userId: string) =>
-						Effect.gen(function* () {
-							const items = yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsDiffTracking)
-									.where(eq(tsDiffTracking.userId, userId))
-									.orderBy(desc(tsDiffTracking.timestamp))
-							);
-
-							return yield* fixDiff(items);
-						}).pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`diffTracking.get.byUserId.all Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-					latest: (userId: string, count: number) =>
-						Effect.gen(function* () {
-							const items = yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsDiffTracking)
-									.where(eq(tsDiffTracking.userId, userId))
-									.orderBy(desc(tsDiffTracking.timestamp))
-							);
-
-							const split = items.slice(0, count);
-
-							return yield* fixDiff(split);
-						}).pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`diffTracking.get.byUserId.latest Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-				},
-				single: (id: string) =>
-					Effect.gen(function* () {
-						const data = yield* dbService.execute((db) =>
-							db.select().from(tsDiffTracking).where(eq(tsDiffTracking.id, id)).get()
-						);
-						if (!data) return;
-						return yield* fixDiff(data);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`diffTracking.get.single Error: ${cause}`),
-									})
-								),
-						})
-					),
-			},
-			revertToDiff: (id: string, type: 'content' | 'data' | 'both') =>
-				Effect.gen(function* () {
-					const diffEntry = yield* dbService.execute((db) =>
-						db.select().from(tsDiffTracking).where(eq(tsDiffTracking.id, id)).get()
-					);
-
-					if (!diffEntry) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error('Diff not found'),
-							})
-						);
-					}
-
-					const shouldRevertData = type === 'data' || type === 'both';
-					const shouldRevertContent = type === 'content' || type === 'both';
-
-					if (shouldRevertData) {
-						const pageData = yield* Effect.try(() => JSON.parse(diffEntry.pageMetaData as string));
-
-						yield* dbService.execute((db) =>
-							db.update(tsPageData).set(pageData.start).where(eq(tsPageData.id, pageData.end.id))
-						);
-					}
-
-					if (shouldRevertContent) {
-						yield* dbService.execute((db) =>
-							db
-								.update(tsPageContent)
-								// @ts-expect-error Drizzle... removed this from the type?
-								.set({ content: diffEntry.pageContentStart })
-								.where(eq(tsPageContent.contentId, diffEntry.pageId))
-						);
-					}
-
-					const allDiffs = yield* dbService.execute((db) =>
-						db
-							.select()
-							.from(tsDiffTracking)
-							.where(eq(tsDiffTracking.pageId, diffEntry.pageId))
-							.orderBy(desc(tsDiffTracking.timestamp))
-					);
-
-					const diffIndex = allDiffs.findIndex((diff) => diff.id === id);
-
-					const diffsToPurge = allDiffs.slice(diffIndex + 1);
-
-					const purgeDiff = dbService.makeQuery((ex, id: string) =>
-						ex((db) => db.delete(tsDiffTracking).where(eq(tsDiffTracking.id, id)))
-					);
-
-					for (const { id } of diffsToPurge) {
-						yield* purgeDiff(id);
-					}
-
-					return yield* fixDiff(diffEntry);
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'LibSQLDatabaseError',
-									cause: new StudioCMS_SDK_Error(`diffTracking.revertToDiff Error: ${cause}`),
-								})
-							),
-						UnknownException: (cause) =>
-							Effect.fail(
-								new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error(`diffTracking.revertToDiff Error: ${cause}`),
-								})
-							),
-					})
-				),
-			utils: {
-				// biome-ignore lint/suspicious/noExplicitAny: <explanation>
-				getMetaDataDifferences: <T extends Record<string, any>>(obj1: T, obj2: T) =>
-					Effect.gen(function* () {
-						// biome-ignore lint/suspicious/noExplicitAny: <explanation>
-						const differences: { label: string; previous: any; current: any }[] = [];
-
-						const Labels: Record<string, string> = {
-							package: 'Page Type',
-							title: 'Page Title',
-							description: 'Page Description',
-							showOnNav: 'Show in Navigation',
-							slug: 'Page Slug',
-							contentLang: 'Content Language',
-							heroImage: 'Hero/OG Image',
-							categories: 'Page Categories',
-							tags: 'Page Tags',
-							showAuthor: 'Show Author',
-							showContributors: 'Show Contributors',
-							parentFolder: 'Parent Folder',
-							draft: 'Draft',
-						};
-
-						const processLabel = (label: string) =>
-							Effect.try(() => (Labels[label] ? Labels[label] : label));
-
-						for (const label in obj1) {
-							const blackListedLabels: string[] = [
-								'publishedAt',
-								'updatedAt',
-								'authorId',
-								'contributorIds',
-							];
-							if (blackListedLabels.includes(label)) continue;
-
-							// biome-ignore lint/suspicious/noPrototypeBuiltins: <explanation>
-							if (obj1.hasOwnProperty(label) && obj2.hasOwnProperty(label)) {
-								if (obj1[label] !== obj2[label]) {
-									if (Array.isArray(obj1[label]) && Array.isArray(obj2[label])) {
-										if (obj1[label].length === obj2[label].length) continue;
-									}
-									differences.push({
-										label: yield* processLabel(label),
-										previous: obj1[label],
-										current: obj2[label],
-									});
-								}
-							}
-						}
-
-						return differences;
-					}).pipe(
-						Effect.catchTags({
-							UnknownException: (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'UNKNOWN',
-										cause: new StudioCMS_SDK_Error(
-											`diffTracking.utils.getMetaDataDifferences Error: ${cause}`
-										),
-									})
-								),
-						})
-					),
-				getDiffHTML: (diff: string | null, options?: Diff2HtmlConfig) =>
-					Effect.try({
-						try: () =>
-							html(diff || '', {
-								diffStyle: 'word',
-								matching: 'lines',
-								drawFileList: false,
-								outputFormat: 'side-by-side',
-								...options,
-							}),
-						catch: (cause) =>
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error(`diffTracking.utils.getDiffHTML Error: ${cause}`),
-							}),
-					}),
-			},
-		};
-
-		const notificationSettings = {
-			site: {
-				get: () =>
-					Effect.gen(function* () {
-						const data = yield* dbService.execute((db) =>
-							db
-								.select()
-								.from(tsNotificationSettings)
-								.where(eq(tsNotificationSettings.id, CMSNotificationSettingsId))
-								.get()
-						);
-
-						if (!data) {
-							return yield* dbService.execute((db) =>
-								db
-									.insert(tsNotificationSettings)
-									.values(NotificationSettingsDefaults)
-									.returning()
-									.get()
-							);
-						}
-
-						return data;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`notificationSettings.site.get Error: ${cause}`),
-									})
-								),
-						})
-					),
-				update: dbService.makeQuery((ex, settings: tsNotificationSettingsInsert) =>
-					ex((db) =>
-						db
-							.update(tsNotificationSettings)
-							.set(settings)
-							.where(eq(tsNotificationSettings.id, CMSNotificationSettingsId))
-							.returning()
-							.get()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(
-											`notificationSettings.site.update Error: ${cause}`
-										),
-									})
-								),
-						})
-					)
-				),
-			},
-		};
-
-		const INIT = {
-			/**
-			 * Initializes the StudioCMS SiteConfig table with the provided configuration.
-			 *
-			 * @param config - The configuration to insert into the SiteConfig table.
-			 * @returns A promise that resolves to the inserted site configuration.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the site configuration.
-			 */
-			siteConfig: (config: Omit<tsSiteConfigSelect, 'id'>) =>
-				dbService
-					.execute((db) =>
-						db
-							.insert(tsSiteConfig)
-							// @ts-expect-error Drizzle... removed this from the type?
-							.values({ ...config, id: CMSSiteConfigId })
-							.returning()
-							.get()
-					)
-					.pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`INIT.siteConfig Error: ${cause}`),
-									})
-								),
-						})
-					),
-			/**
-			 * Initializes the StudioCMS Ghost User.
-			 *
-			 * The ghost user is a default user that is used to perform actions on behalf of the system as well as to replace deleted users.
-			 *
-			 * @returns A promise that resolves to the ghost user record.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the ghost user.
-			 */
-			ghostUser: () =>
-				Effect.gen(function* () {
-					const ghostUser = yield* AUTH.user.ghost.verifyExists();
-					if (!ghostUser) return yield* AUTH.user.ghost.create();
-					const ghostUserRecord = yield* AUTH.user.ghost.get();
-					if (!ghostUserRecord) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'LibSQLDatabaseError',
-								cause: new StudioCMS_SDK_Error(
-									'INIT.ghostUser Error: Error getting ghost user from database: The ghost user may not exist yet.'
-								),
-							})
-						);
-					}
-					return ghostUserRecord;
-				}),
-		};
-
-		const AUTH = {
-			verifyEmail: {
-				get: dbService.makeQuery((ex, id: string) =>
-					ex((db) =>
-						db
-							.select()
-							.from(tsEmailVerificationTokens)
-							.where(eq(tsEmailVerificationTokens.id, id))
-							.get()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.verifyEmail.get Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-				create: (userId: string) =>
-					Effect.gen(function* () {
-						yield* dbService.execute((db) =>
-							db
-								.delete(tsEmailVerificationTokens)
-								.where(eq(tsEmailVerificationTokens.userId, userId))
-						);
-
-						const token = yield* generateToken(userId);
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsEmailVerificationTokens)
-								.values({
-									// @ts-expect-error Drizzle... removed this from the type?
-									id: crypto.randomUUID(),
-									userId,
-									token,
-									expiresAt: new Date(Date.now() + 1000 * 60 * 10),
-								})
-								.returning()
-								.get()
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.verifyEmail.create Error: ${cause}`),
-									})
-								),
-						})
-					),
-				delete: dbService.makeQuery((ex, userId: string) =>
-					ex((db) =>
-						db.delete(tsEmailVerificationTokens).where(eq(tsEmailVerificationTokens.userId, userId))
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.verifyEmail.delete Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-			},
-			oAuth: {
-				/**
-				 * Creates a new OAuth account in the database.
-				 *
-				 * @param data - The data to insert into the OAuth account table.
-				 * @returns A promise that resolves to the inserted OAuth account.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the OAuth account.
-				 */
-				create: dbService.makeQuery((ex, data: tsOAuthAccountsSelect) =>
-					ex((db) => db.insert(tsOAuthAccounts).values(data).returning().get()).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.oAuth.create Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-				/**
-				 * Deletes an OAuth account from the database.
-				 *
-				 * @param userId - The ID of the user associated with the OAuth account.
-				 * @param provider - The provider of the OAuth account.
-				 * @returns A promise that resolves to a deletion response.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the OAuth account.
-				 */
-				delete: (userId: string, provider: string) =>
-					Effect.gen(function* () {
-						yield* dbService.execute((db) =>
-							db
-								.delete(tsOAuthAccounts)
-								.where(
-									and(eq(tsOAuthAccounts.userId, userId), eq(tsOAuthAccounts.provider, provider))
-								)
-						);
-
-						return {
-							status: 'success',
-							message: 'OAuth account deleted',
-						};
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.oAuth.delete Error: ${cause}`),
-									})
-								),
-						})
-					),
-				/**
-				 * Searches for OAuth accounts based on the provider ID and user ID.
-				 *
-				 * @param providerId - The provider ID to search for.
-				 * @param userId - The user ID to search for.
-				 * @returns A promise that resolves to the OAuth account data if found, otherwise undefined.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while searching for the OAuth account.
-				 */
-				searchProvidersForId: (providerId: string, userId: string) =>
-					dbService
-						.execute((db) =>
-							db
-								.select()
-								.from(tsOAuthAccounts)
-								.where(
-									and(
-										eq(tsOAuthAccounts.providerUserId, providerId),
-										eq(tsOAuthAccounts.userId, userId)
-									)
-								)
-								.get()
-						)
-						.pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`AUTH.oAuth.searchProviderForId Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-			},
-			permission: {
-				/**
-				 * Checks the current status of a user's permissions.
-				 */
-				currentStatus: dbService.makeQuery((ex, userId: string) =>
-					ex((db) =>
-						db.select().from(tsPermissions).where(eq(tsPermissions.user, userId)).get()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.permission.currentStatus Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-			},
-			session: {
-				/**
-				 * Creates a new session in the database.
-				 *
-				 * @param data - The data to insert into the session table.
-				 * @returns A promise that resolves to the inserted session.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the session.
-				 */
-				create: dbService.makeQuery((ex, data: tsSessionTableInsert) =>
-					ex((db) =>
-						db
-							.insert(tsSessionTable)
-							.values(data)
-							.returning({
-								id: tsSessionTable.id,
-								userId: tsSessionTable.userId,
-								expiresAt: tsSessionTable.expiresAt,
-							})
-							.get()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.session.create Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-				/**
-				 * Gets a session with the associated user.
-				 *
-				 * @param sessionId - The ID of the session to search for.
-				 * @returns A promise that resolves to the session with the associated user.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the session with the user.
-				 */
-				sessionWithUser: dbService.makeQuery((ex, sessionId: string) =>
-					ex((db) =>
-						db
-							.select({ user: tsUsers, session: tsSessionTable })
-							.from(tsSessionTable)
-							.innerJoin(tsUsers, eq(tsSessionTable.userId, tsUsers.id))
-							.where(eq(tsSessionTable.id, sessionId))
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.session.sessionWithUser Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-				/**
-				 * Deletes a session from the database.
-				 *
-				 * @param sessionId - The ID of the session to delete.
-				 * @returns A promise that resolves to a deletion response.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the session.
-				 */
-				delete: dbService.makeQuery((ex, sessionId: string) =>
-					ex((db) =>
-						db
-							.delete(tsSessionTable)
-							.where(eq(tsSessionTable.id, sessionId))
-							.then(() => ({
-								status: 'success',
-								message: 'Session deleted',
-							}))
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.session.delete Error: ${cause}`),
-									})
-								),
-						})
-					)
-				),
-				/**
-				 * Updates the expiration date of a session.
-				 *
-				 * @param sessionId - The ID of the session to update.
-				 * @param newDate - The new expiration date for the session.
-				 * @returns A promise that resolves to the updated session.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the session.
-				 */
-				update: (sessionId: string, newDate: Date) =>
-					dbService
-						.execute((db) =>
-							db
-								.update(tsSessionTable)
-								.set({ expiresAt: newDate })
-								.where(eq(tsSessionTable.id, sessionId))
-								.returning()
-						)
-						.pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(`AUTH.session.update Error: ${cause}`),
-										})
-									),
-							})
-						),
-			},
-			user: {
-				/**
-				 * Creates a new user in the database.
-				 *
-				 * @param newUserData - The data to insert into the users table.
-				 * @returns A promise that resolves to the inserted user.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the user.
-				 */
-				create: (newUserData: tsUsersInsert, rank?: 'visitor' | 'editor' | 'admin' | 'owner') =>
-					Effect.gen(function* () {
-						const newUser = yield* dbService.execute((db) =>
-							db.insert(tsUsers).values(newUserData).returning().get()
-						);
-						yield* dbService.execute((db) =>
-							db.insert(tsPermissions).values({ user: newUser.id, rank: rank || 'visitor' })
-						);
-						return newUser;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(`AUTH.user.create Error: ${cause}`),
-									})
-								),
-						})
-					),
-				/**
-				 * Updates a user in the database.
-				 *
-				 * @param userId - The ID of the user to update.
-				 * @param userData - The data to update the user with.
-				 * @returns A promise that resolves to the updated user.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the user.
-				 */
-				update: (userId: string, userData: tsUsersUpdate) =>
-					dbService
-						.execute((db) =>
-							db.update(tsUsers).set(userData).where(eq(tsUsers.id, userId)).returning().get()
-						)
-						.pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(`AUTH.user.update Error: ${cause}`),
-										})
-									),
-							})
-						),
-				/**
-				 * Searches for users based on the provided username or email.
-				 *
-				 * @param username - The username to search for.
-				 * @param email - The email to search for.
-				 * @returns A promise that resolves to an object containing the search results for the username and email.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while searching for the username or email.
-				 */
-				searchUsersForUsernameOrEmail: (username = '', email = '') =>
-					Effect.gen(function* () {
-						const usernameSearch = yield* dbService.execute((db) =>
-							db.select().from(tsUsers).where(eq(tsUsers.username, username))
-						);
-						const emailSearch = yield* dbService.execute((db) =>
-							db.select().from(tsUsers).where(eq(tsUsers.email, email))
-						);
-
-						return { usernameSearch, emailSearch };
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(
-											`AUTH.user.searchUsersForUsernameOrEmail Error: ${cause}`
-										),
-									})
-								),
-						})
-					),
-				ghost: {
-					/**
-					 * Verifies if the ghost user exists in the database.
-					 *
-					 * @returns A promise that resolves to a boolean indicating if the ghost user exists.
-					 * @throws {StudioCMS_SDK_Error} If an error occurs while verifying the ghost user.
-					 */
-					verifyExists: () =>
-						Effect.gen(function* () {
-							const ghostUser = yield* dbService.execute((db) =>
-								db.select().from(tsUsers).where(eq(tsUsers.id, GhostUserDefaults.id)).get()
-							);
-							if (!ghostUser) return false;
-							return true;
-						}).pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									Effect.fail(
-										new SDKCoreError({
-											type: 'LibSQLDatabaseError',
-											cause: new StudioCMS_SDK_Error(
-												`AUTH.user.ghost.verifyExists Error: ${cause}`
-											),
-										})
-									),
-							})
-						),
-					/**
-					 * Creates the ghost user in the database.
-					 *
-					 * @returns A promise that resolves to the inserted ghost user.
-					 * @throws {StudioCMS_SDK_Error} If an error occurs while creating the ghost user.
-					 */
-					create: () =>
-						dbService
-							.execute((db) => db.insert(tsUsers).values(GhostUserDefaults).returning().get())
-							.pipe(
-								Effect.catchTags({
-									'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-										Effect.fail(
-											new SDKCoreError({
-												type: 'LibSQLDatabaseError',
-												cause: new StudioCMS_SDK_Error(`AUTH.user.ghost.create Error: ${cause}`),
-											})
-										),
-								})
-							),
-					/**
-					 * Gets the ghost user from the database.
-					 *
-					 * @returns A promise that resolves to the ghost user.
-					 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the ghost user.
-					 */
-					get: () =>
-						dbService
-							.execute((db) =>
-								db.select().from(tsUsers).where(eq(tsUsers.id, GhostUserDefaults.id)).get()
-							)
-							.pipe(
-								Effect.catchTags({
-									'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-										Effect.fail(
-											new SDKCoreError({
-												type: 'LibSQLDatabaseError',
-												cause: new StudioCMS_SDK_Error(`AUTH.user.ghost.get Error: ${cause}`),
-											})
-										),
-								})
-							),
-				},
-			},
-		};
 
 		return {
 			db,
@@ -1185,20 +151,27 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 			collectTags,
 			collectPageData,
 			collectUserData,
-			CLEAR, // migrated
-			DELETE, // migrated
-			REST_API, // migrated
-			POST, // migrated
-			UPDATE, // migrated
-			GET, // migrated
-			INIT,
-			AUTH,
 			resetTokenBucket,
 			diffTracking,
 			notificationSettings,
+			CLEAR,
+			DELETE,
+			REST_API,
+			POST,
+			UPDATE,
+			GET,
+			AUTH,
+			INIT,
 		};
 	}),
 }) {
 	static Provide = Effect.provide(this.Default);
-	static Cache = CacheContext.makeProvide({ pages, FolderList, folderTree, pageFolderTree, siteConfig, version })
+	static Cache = CacheContext.makeProvide({
+		pages,
+		FolderList,
+		folderTree,
+		pageFolderTree,
+		siteConfig,
+		version,
+	});
 }
